Drop PropertyMapper, Drop JsonEvaluatorInterface
